requirements.txt
python-docx
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
lxml